+++ tests/strategies/test_deepspeed_strategy.py
@@ -276,9 +276,9 @@ def test_deepspeed_run_configure_optimizers(tmpdir):
 
             assert isinstance(trainer.optimizers[0], FP16_DeepSpeedZeroOptimizer)
             assert isinstance(trainer.optimizers[0].optimizer, torch.optim.SGD)
-            assert isinstance(trainer.lr_schedulers[0]["scheduler"], torch.optim.lr_scheduler.StepLR)
+            assert isinstance(trainer.lr_scheduler_configs[0].scheduler, torch.optim.lr_scheduler.StepLR)
             # check that the lr_scheduler config was preserved
-            assert trainer.lr_schedulers[0]["name"] == "Sean"
+            assert trainer.lr_scheduler_configs[0].name == "Sean"
 
     class TestModel(BoringModel):
         def configure_optimizers(self):
@@ -314,7 +314,7 @@ def test_deepspeed_config(tmpdir, deepspeed_zero_config):
 
             assert isinstance(trainer.optimizers[0], FP16_DeepSpeedZeroOptimizer)
             assert isinstance(trainer.optimizers[0].optimizer, torch.optim.SGD)
-            assert isinstance(trainer.lr_schedulers[0]["scheduler"], WarmupLR)
+            assert isinstance(trainer.lr_scheduler_configs[0].scheduler, WarmupLR)
 
     model = BoringModel()
     trainer = Trainer(
@@ -716,8 +716,8 @@ def test_deepspeed_multigpu_stage_3_resume_training(tmpdir):
             assert trainer.current_epoch == 1
 
             # assert lr-scheduler states are loaded correctly
-            original_lr_scheduler = initial_trainer.lr_schedulers[0]["scheduler"]
-            current_lr_scheduler = trainer.lr_schedulers[0]["scheduler"]
+            original_lr_scheduler = initial_trainer.lr_scheduler_configs[0].scheduler
+            current_lr_scheduler = trainer.lr_scheduler_configs[0].scheduler
             assert original_lr_scheduler.state_dict() == current_lr_scheduler.state_dict()
 
     model = ModelParallelClassificationModel()